// src/types/receipts.ts
export interface Receipt {
    receiptNumber: number;          // 8371
    receiptDate: string;            // 2025-01-22
    receiptReference: string;       // 2ND QTR 2024 - PROP TREATY
    brokerCedantName: string;       // HERITAGE INSURANCE CO LTD
    address1: string;               // 4th Floor, Bains Avenue
    address2: string;               // Masaki Icon, Msasani Pensular
    address3: string;               // P.O.BOX:7390, DAR ES SALAAM
    address4: string;               // TANZANIA
    instrumentType: string;         // B
    instrumentNumber: string;       // TISS
    instrumentDate: string;         // 2025-01-20
    instrumentBankName: string;     // CRDB
    instrumentCurrencyCode: string; // TZS
    instrumentAmount: string;       // 62,149,638.050
}